+++ src/crypter.rs
@@ -56,6 +56,9 @@ use bytes::{Buf, BytesMut};
 use sodiumoxide::crypto::secretstream::{self, Header, Key, Pull, Push, Stream, Tag};
 use std::{cmp, fmt, io::Write, mem};
 
+#[cfg(feature = "io-async")]
+pub use crate::async_::crypter::*;
+
 /// Minimum block size allowed - values smaller than this will automatically be
 /// coerced up to this value.
 pub const MIN_BLOCK_SIZE: usize = 1024;
@@ -327,7 +330,7 @@ impl StateMachine for Encrypter {
 type KeyLookupFn = Box<dyn FnOnce(&PublicKey) -> Option<SecretKey>>;
 
 #[derive(strum_macros::AsRefStr)]
-pub(crate) enum KeyResolution {
+enum KeyResolution {
     Available(PublicKey, SecretKey),
     Deferred(KeyLookupFn),
 }
@@ -340,10 +343,10 @@ impl fmt::Debug for KeyResolution {
 
 #[derive(strum_macros::AsRefStr)]
 pub(crate) enum DecrypterState {
-    ReadPreheader(KeyResolution),
-    ReadPublicKey(KeyResolution),
-    SecretKeyLookup(PublicKey, KeyLookupFn),
-    ReadHeader(PublicKey, PublicKey, SecretKey),
+    ReadPreheader,
+    ReadPublicKey,
+    SecretKeyLookup(PublicKey),
+    ReadHeader(PublicKey, SecretKey),
     OpenStream(Key, Header),
     ReadLength(Stream<Pull>),
     ReadBlock(Stream<Pull>, usize),
@@ -358,14 +361,18 @@ impl fmt::Debug for DecrypterState {
     }
 }
 
+/// Result of update call (read, written) or require secret key to make
+/// progress.
+pub(crate) enum UpdateReturn {
+    Progress(usize, usize),
+    NeedSecretKey(usize, usize, PublicKey),
+}
+
 /// Low-level interface to decrypting data in the saltlick format.
 #[derive(Debug)]
 pub struct Decrypter {
-    ciphertext: BytesMut,
-    consumed: usize,
-    last_block_size: Option<usize>,
-    plaintext: Vec<u8>,
-    state: Option<DecrypterState>,
+    inner: DecrypterInner,
+    key_resolution: Option<KeyResolution>,
 }
 
 impl Decrypter {
@@ -373,11 +380,8 @@ impl Decrypter {
     pub fn new(public_key: PublicKey, secret_key: SecretKey) -> Decrypter {
         let key = KeyResolution::Available(public_key, secret_key);
         Decrypter {
-            ciphertext: BytesMut::new(),
-            consumed: 0,
-            last_block_size: None,
-            plaintext: Vec::new(),
-            state: Some(DecrypterState::ReadPreheader(key)),
+            inner: DecrypterInner::default(),
+            key_resolution: Some(key),
         }
     }
 
@@ -396,11 +400,8 @@ impl Decrypter {
     {
         let key = KeyResolution::Deferred(Box::new(lookup_fn));
         Decrypter {
-            ciphertext: BytesMut::new(),
-            consumed: 0,
-            last_block_size: None,
-            plaintext: Vec::new(),
-            state: Some(DecrypterState::ReadPreheader(key)),
+            inner: DecrypterInner::default(),
+            key_resolution: Some(key),
         }
     }
 
@@ -414,83 +415,145 @@ impl Decrypter {
         mut input: &[u8],
         mut output: &mut [u8],
     ) -> Result<(usize, usize), SaltlickError> {
+        match self.inner.update(input, &mut output, None)? {
+            UpdateReturn::Progress(nread, nwritten) => Ok((nread, nwritten)),
+            UpdateReturn::NeedSecretKey(nread, nwritten, public_key) => {
+                let secret = match self.key_resolution.take() {
+                    Some(KeyResolution::Available(public, secret)) => {
+                        if public == public_key {
+                            Ok(secret)
+                        } else {
+                            Err(SaltlickError::PublicKeyMismatch)
+                        }
+                    }
+                    Some(KeyResolution::Deferred(lookup_fn)) => {
+                        lookup_fn(&public_key).ok_or(SaltlickError::SecretKeyNotFound)
+                    }
+                    None => Err(SaltlickError::SecretKeyNotFound),
+                }?;
+                advance_slice(&mut input, nread);
+                advance_slice_mut(&mut output, nwritten);
+                match self.inner.update(input, &mut output, Some(secret))? {
+                    UpdateReturn::Progress(read, written) => Ok((nread + read, nwritten + written)),
+                    UpdateReturn::NeedSecretKey(_, _, _) => unreachable!(),
+                }
+            }
+        }
+    }
+
+    /// Convenience version of `update` that allocates and returns output data
+    /// as a `Vec<u8>`.
+    pub fn update_to_vec(&mut self, input: impl AsRef<[u8]>) -> Result<Vec<u8>, SaltlickError> {
+        let input = input.as_ref();
+        let mut plaintext = vec![0u8; self.inner.estimate_output_size(input.len())];
+        let (rd, wr) = self.update(input, &mut plaintext)?;
+
+        // The decrypter never buffers more than 1 complete block, and input
+        // data will never decrypt larger than the ciphertext size. Since we
+        // allocate space for 2 complete blocks plus the input size, we will
+        // always consume the entire input. If we don't, an assumption has
+        // broken down and the output is not valid.
+        assert!(rd == input.len());
+
+        plaintext.truncate(wr);
+        Ok(plaintext)
+    }
+
+    /// Returns true if the crypter has been finalized.
+    pub fn is_finalized(&self) -> bool {
+        self.inner.is_finalized()
+    }
+}
+
+/// Inner decryption state machine.
+///
+/// The inner logic that performs stream decryption in a non-blocking fashion -
+/// this type will always return as soon as all available data has been
+/// processed. Wrapper types make use of this inner type to provide unique
+/// key-lookup implementations.
+#[derive(Debug)]
+pub(crate) struct DecrypterInner {
+    ciphertext: BytesMut,
+    consumed: usize,
+    last_block_size: Option<usize>,
+    plaintext: Vec<u8>,
+    state: Option<DecrypterState>,
+}
+
+impl DecrypterInner {
+    pub fn update(
+        &mut self,
+        mut input: &[u8],
+        mut output: &mut [u8],
+        secret_key: Option<SecretKey>,
+    ) -> Result<UpdateReturn, SaltlickError> {
         use self::DecrypterState::*;
         let mut nread = 0;
         let mut nwritten = 0;
+        use UpdateReturn::{NeedSecretKey, Progress};
         state::turn(self, |next, self_| match next {
-            ReadPreheader(key_resolution) => match read::preheader(&self_.ciphertext)? {
+            ReadPreheader => match read::preheader(&self_.ciphertext)? {
                 ReadStatus::Complete(version, n) => {
                     self_.ciphertext.advance(n);
                     if version != Version::V1 {
                         state::err(SaltlickError::UnsupportedVersion)
                     } else {
-                        state::next(ReadPublicKey(key_resolution))
+                        state::next(ReadPublicKey)
                     }
                 }
                 ReadStatus::Incomplete(_needed) if input.is_empty() => {
-                    state::ret((nread, nwritten), ReadPreheader(key_resolution))
+                    state::ret(Progress(nread, nwritten), ReadPreheader)
                 }
                 ReadStatus::Incomplete(needed) => {
                     let take = cmp::min(needed, input.len());
                     self_.ciphertext.extend_from_slice(&input[..take]);
                     advance_slice(&mut input, take);
                     nread += take;
-                    state::next(ReadPreheader(key_resolution))
+                    state::next(ReadPreheader)
                 }
             },
-            ReadPublicKey(key_resolution) => match read::header_v1_public_key(&self_.ciphertext)? {
+            ReadPublicKey => match read::header_v1_public_key(&self_.ciphertext)? {
                 ReadStatus::Complete(file_public_key, n) => {
                     self_.ciphertext.advance(n);
-                    match key_resolution {
-                        KeyResolution::Available(public_key, secret_key) => {
-                            state::next(ReadHeader(file_public_key, public_key, secret_key))
-                        }
-                        KeyResolution::Deferred(lookup_fn) => {
-                            state::next(SecretKeyLookup(file_public_key, lookup_fn))
-                        }
-                    }
+                    state::ret(
+                        NeedSecretKey(nread, nwritten, file_public_key.clone()),
+                        SecretKeyLookup(file_public_key),
+                    )
                 }
                 ReadStatus::Incomplete(_needed) if input.is_empty() => {
-                    state::ret((nread, nwritten), ReadPublicKey(key_resolution))
+                    state::ret(Progress(nread, nwritten), ReadPublicKey)
                 }
                 ReadStatus::Incomplete(needed) => {
                     let take = cmp::min(needed, input.len());
                     self_.ciphertext.extend_from_slice(&input[..take]);
                     advance_slice(&mut input, take);
                     nread += take;
-                    state::next(ReadPublicKey(key_resolution))
+                    state::next(ReadPublicKey)
                 }
             },
-            SecretKeyLookup(file_public_key, lookup_fn) => {
-                if let Some(secret_key) = lookup_fn(&file_public_key) {
-                    state::next(ReadHeader(
-                        file_public_key.clone(),
-                        file_public_key,
-                        secret_key,
-                    ))
+            SecretKeyLookup(file_public_key) => {
+                if let Some(secret) = secret_key.clone() {
+                    state::next(ReadHeader(file_public_key, secret))
                 } else {
                     state::err(SaltlickError::SecretKeyNotFound)
                 }
             }
-            ReadHeader(file_public_key, public_key, secret_key) => {
-                if file_public_key != public_key {
-                    return state::err(SaltlickError::PublicKeyMismatch);
-                }
+            ReadHeader(public_key, secret_key) => {
                 match read::header_v1_sealed_text(&self_.ciphertext, &public_key, &secret_key)? {
                     ReadStatus::Complete((key, header), n) => {
                         self_.ciphertext.advance(n);
                         state::next(OpenStream(key, header))
                     }
                     ReadStatus::Incomplete(_needed) if input.is_empty() => state::ret(
-                        (nread, nwritten),
-                        ReadHeader(file_public_key, public_key, secret_key),
+                        Progress(nread, nwritten),
+                        ReadHeader(public_key, secret_key),
                     ),
                     ReadStatus::Incomplete(needed) => {
                         let take = cmp::min(needed, input.len());
                         self_.ciphertext.extend_from_slice(&input[..take]);
                         advance_slice(&mut input, take);
                         nread += take;
-                        state::next(ReadHeader(file_public_key, public_key, secret_key))
+                        state::next(ReadHeader(public_key, secret_key))
                     }
                 }
             }
@@ -505,7 +568,7 @@ impl Decrypter {
                     state::next(ReadBlock(stream, length))
                 }
                 ReadStatus::Incomplete(_needed) if input.is_empty() => {
-                    state::ret((nread, nwritten), ReadLength(stream))
+                    state::ret(Progress(nread, nwritten), ReadLength(stream))
                 }
                 ReadStatus::Incomplete(needed) => {
                     let take = cmp::min(needed, input.len());
@@ -523,7 +586,7 @@ impl Decrypter {
                         state::next(FlushOutput(stream, finalized))
                     }
                     ReadStatus::Incomplete(_needed) if input.is_empty() => {
-                        state::ret((nread, nwritten), ReadBlock(stream, length))
+                        state::ret(Progress(nread, nwritten), ReadBlock(stream, length))
                     }
                     ReadStatus::Incomplete(needed) => {
                         let take = cmp::min(needed, input.len());
@@ -542,7 +605,7 @@ impl Decrypter {
                         state::next(ReadLength(stream))
                     }
                 } else if output.is_empty() {
-                    state::ret((nread, nwritten), FlushOutput(stream, finalized))
+                    state::ret(Progress(nread, nwritten), FlushOutput(stream, finalized))
                 } else {
                     let take = cmp::min(output.len(), self_.plaintext_len());
                     let n = output
@@ -553,30 +616,11 @@ impl Decrypter {
                     state::next(FlushOutput(stream, finalized))
                 }
             }
-            Finalized => state::ret((nread, nwritten), Finalized),
+            Finalized => state::ret(Progress(nread, nwritten), Finalized),
             Errored => state::err(SaltlickError::StateMachineErrored),
         })
     }
 
-    /// Convenience version of `update` that allocates and returns output data
-    /// as a `Vec<u8>`.
-    pub fn update_to_vec(&mut self, input: impl AsRef<[u8]>) -> Result<Vec<u8>, SaltlickError> {
-        let input = input.as_ref();
-        let mut plaintext = vec![0u8; self.estimate_output_size(input.len())];
-        let (rd, wr) = self.update(input, &mut plaintext)?;
-
-        // The decrypter never buffers more than 1 complete block, and input
-        // data will never decrypt larger than the ciphertext size. Since we
-        // allocate space for 2 complete blocks plus the input size, we will
-        // always consume the entire input. If we don't, an assumption has
-        // broken down and the output is not valid.
-        assert!(rd == input.len());
-
-        plaintext.truncate(wr);
-        Ok(plaintext)
-    }
-
-    /// Returns true if the crypter has been finalized.
     pub fn is_finalized(&self) -> bool {
         match self.state {
             Some(DecrypterState::Finalized) => true,
@@ -584,7 +628,7 @@ impl Decrypter {
         }
     }
 
-    fn estimate_output_size(&self, input_len: usize) -> usize {
+    pub fn estimate_output_size(&self, input_len: usize) -> usize {
         input_len + (2 * self.last_block_size.unwrap_or(DEFAULT_BLOCK_SIZE))
     }
 
@@ -593,9 +637,21 @@ impl Decrypter {
     }
 }
 
-impl StateMachine for Decrypter {
+impl Default for DecrypterInner {
+    fn default() -> DecrypterInner {
+        DecrypterInner {
+            ciphertext: BytesMut::new(),
+            consumed: 0,
+            last_block_size: None,
+            plaintext: Vec::new(),
+            state: Some(DecrypterState::ReadPreheader),
+        }
+    }
+}
+
+impl StateMachine for DecrypterInner {
     type State = DecrypterState;
-    type Return = (usize, usize);
+    type Return = UpdateReturn;
     type Error = SaltlickError;
 
     fn take_state(&mut self) -> Self::State {
@@ -616,12 +672,12 @@ impl StateMachine for Decrypter {
 // we've already written. This is safe because a slice is always a reference to
 // some other data - all we're doing is updating the pointer so it points to a
 // new subset of that same data.
-fn advance_slice<T>(slice: &mut &[T], n: usize) {
+pub(crate) fn advance_slice<T>(slice: &mut &[T], n: usize) {
     let (_a, b) = mem::replace(slice, &[]).split_at(n);
     *slice = b;
 }
 
-fn advance_slice_mut<T>(slice: &mut &mut [T], n: usize) {
+pub(crate) fn advance_slice_mut<T>(slice: &mut &mut [T], n: usize) {
     let (_a, b) = mem::replace(slice, &mut []).split_at_mut(n);
     *slice = b;
 }
@@ -832,7 +888,6 @@ mod tests {
     fn deferred_key_load_test() {
         let test_data = random_bytes(4, 25000);
         let (public, secret) = key::gen_keypair();
-
         let mut encrypter = Encrypter::new(public);
         let ciphertext = encrypter.update_to_vec(&test_data[..], true).unwrap();